Silence several warnings
Added many of the warning flags from Mantid, which highlighted some
issues in the ADARA packet and parsing code. EPICS headers have many
warnings, so we added a test program to give us a place to build only
the ADARA code with the hightened warning level. This lets us vet
changes before they show up in Mantid but avoid the spew of EPICS.