Process raw events into banked events
DataSource is responsible for detecting the boundaries of each pulse and
checking for duplicate pulse identifiers. This is easily done on
reception of the RTDL packet, which contains the charge and veto info
for the previous pulse and the parameters for the new pulse. If we don't
see an RTDL packet, we can still detect them upon receiving new raw
event data.

SMSControl will accumulate the raw events and parse them into
appropriate buckets for the banked event packets (and eventually the
fast meta data and beam monitor information). Once all sources for a
given pulse have marked the it complete, SMSControl will walk the list
of pending pulses and complete each pulse that comes before the most
recently completed pulse. This catches cases where one preprocessor has
missed the RTDL packet and others have not, causing a duplicate pulse
that would never complete otherwise.