Add hooks to record state at beginning of each data file
SMS needs to record pieces of the current state of the system at the
beginning of each data file to allow recovery in the event that portions
(files) of an experiment are lost.

We add a callback hook so objects that are interested can be called when
a new file is created, prior to event data being stored into the new
file. This will be used to store mapping packets, device descriptions,
and environment state, among other things.

This functionality will be extended in the future to caputre snapshots
of state for newly connected live clients, as well as preserving state
information for historical data replay.