Add real pixel map support
Read in the pixel map from a file and create the packet for the ADARA
stream. We use STL maps to during the construction to validate the
mapping meets our rules (no duplicate values, one-to-one mapping, etc.)
and then transfer the data to a vector for faster lookup. Using a map
also makes it easier to iterate over the mapping when creating the
packet for later use.