Allow creation of files outside of a container
SMS needs to be able to create a StorageFile that does not live in a
given container. We will need this capability to create an index into
the data store to support requests for historical data upon connection,
and we need to be able to create temporary files to store a snapshot of
the system state for when clients want to start at the next pulse.