Send clients a snapshot of the system state on connection
Upon initial connection, the client needs to see a snapshot of the
system state before receiving pulse data. Call into the StorageManager
to generate the snapshot using the existing data dump at the beginning
of each new storage file.  Because we may be joining a stream in
progress, we need to have each file know it's current offset -- the
state file would start at 0, but the event data would start later
in a different file.

This will eventually be extended to allow replay from an indexed
position.