Add support for metadata packets
With this set of changes, we can now receive metadata packets from data
sources. We keep track of the device descriptors we see, and hang the
most recent variable updates off of those. We snapshot the metadata
state when requested by the StorageManager, so that new clients and new
files get the appropriate prologue.

Dealing with devices that go away has not yet been addressed, nor has
generating fast metadata updates.