Use new layout for the pixel map packet
Previously, the pixel map packet listed the logical id and bank for each
physical pixel in the system. While this was sufficient to regenerate
the mapping, the bank information was largely redundant. By putting the
bank in the section header, changing to listing the physical pixels in
the map entries, and implying the logical id for each entry, we can save
space on the wire and in each file containing ADARA data. For CNMS, this
saves approximately 100 KB.