Code review fixes
Dale noticed a few errors -- in the ADARA Parser, we'd leave a
payload-less packet in the buffer until another packet came in, and we
grew the receive buffer faster than intended.

In SMSControl.cc, the programmer would get warning if they added two PVs
with the same name, which is a logic error and should never happen.
Also, the semantics of SMSControl::setRecording() were unclear, so
improve the comment.