Update banked event packet to include source info
We were dropping timing information on the floor with the banked event
packet format. Update the packet to include sections per hardware
source, which will caputre the timing information.

Along the way, we had the sense of bit 31 of the TOF Offset field
inverted; it is set if the TOF Offset has been applied to the events in
the packet. We similarly call out a few fields as u32 that had been
previously implied.

We remove the Ring Period field from the Banked Event and Beam Monitor
packets; this information is redundant of the Pulse Energy field, and
can be obtained from the RTDL packet.

We add addtional flags for the Banked Event and Beam Monitor packets to
capture addtional error cases.

We define that bank ID 0xfffffffe is used for pixels with the error bit
set from the pre-processor, and bank ID 0xffffffff is used for pixels
that have no entry in the pixel mapping table.