Update event handling to new Banked Event format
Begin sending te new banked event packet format, which identifies the
hardware source for each set of events, and the timing information
associated with them. This requires some reworking of the accumulation
code, as we need to bin the data by source.

Along the way, we fixed the inverted meaning of a bit in the TOF offset
field (set did indicate RAW event times, now it indicates corrected).
The banked event and beam monitor packets also removed the ring period
field.

Add decoders to adara-parser so that we can check the output.