Update Beam Monitor packet to contain timing info
As with the old Banked Event packet format, we were dropping timing
information on the floor. Rather than use the approach used for the
banked events, we just move the timing info into each monitor section.
This works, as we expect to have a handfull of monitors on each
beamline, keeping the duplicated information small.