Silence EPICS library warnings
The EPICS gdd code supporting CA servers prints to stderr when it
receives a request type it doesn't know about. To avoid having it spew,
register all of the types we know about, even if we don't use them. We
can return "bad type", and it appears that clients do the right thing,
without the spew.