Add 'Source List' packet
Allow data sources to declare the possible source IDs they will send in
the Raw Event packets. This allows the SMS to notice when we are only
receiving partial data for an event. If the data source will not be
sending Raw Event packets, it should send a Source list packet with no
IDs listed.