Maintain control over container timestamps
Naming non-run containers based on the timestamp of the first packet
written to them leaves us open to bugs in the data sources. After seeing
many issues with upstream timestamps, and given that our scanning of the
data store on startup will depend on accurate timestamps, it is much
better to just control them ourselves.