Don't pound on STS if we're getting transient errors
We may see issues form the STS such as memory pressure and HDF5 spitting
out error messages over the socket. When we see these problems, we
should not continuously pound on its port, as that just fills our logs
with useless messages. Instead, back off for a configurable time --
currently 60 seconds -- before we re-open the floodgates.