Restructure repo as a common base
We want to have a single repo to collect the ADARA code. Since a few
tools use SMS's parser, it seems like a good candidate to be the repo
we merge everything into.

Split out the code into common, tool, and test directories, and make
the Makefile.am more modular.