SMS: inform STSClientMgr of pending runs from startup scan
STSClientMgr becomes a singleton object so that StorageManager can tell
it about the runs it found during the startup scan. This has the
unfortunate side effect of adding additional depenencies to the small
test/example code.

We also need to tell STSClientMgr to start processing the queue after
we've told it about the pending runs we found during the startup scan,
as they will otherwise sit idle until we start recording our first run.