SMS: be more resilient to missing files
Rather than kill ourselves trying to make sure that we never encounter
the rare corner case where we purge files that we are currently sending
to a client, just cleanly kill off the client and continue processing
events for the rest of the system.

Similarly, we should handle missing files gracefully when sending runs
to STS, though this situation is a serious error, and will need to be
delt with manually.