SMS: don't delete the current container
If we the amount of unpurgable (ie, untranslated) experiment data is
over the SMS's desired storage limit, we could try to delete the current
container if it is not a run. This would lead to a crash when we try to
create the next file in that container.

This is an unlikely scenario, but it is easy to avoid: don't delete the
directory housing the last container in the last daily directory. This
may not always be the current container, as the current daily directory
may not yet be in the cache for purging, but we'll clean up any leftover
directories once we repopulate the cache to include the current day.