SMS: shutdown STS socket when we've send all of our data
If SMS died in the middle of a recording, we'll try to resend the data
to STS when we restart. STS will hang around forever waiting for the
final RunStatus packet to tell it we've finished, but it will never
come.

Have SMS shutdown the write side of the socket -- this will give STS and
EOF indication, so it knows no more data is coming, and we will mark the
run for manual processing.