SMS: create new file after notifying about new container
Commit c9628ffb (SMS: Always create a file for an experimental run)
fixed a problem with lingering connections to STS, but broke live
clients when new runs are started. The change to have the container
create the new file in StorageContainer::create() causes the new file to
be created before interested parties are notified of the new container.
This prevents them from signing up for notifications of new files before
the initial file is created. Live clients would not receive any data
until we begin the second file of the recording.

We fix this by moving responsibility for the initial file creation back
to StorageManager.