SMS: make logic errors fatal
As much of the work of SMS is performed in response to new data arriving
in the DataSource class, the effect of run time errors is limited to
closing the socket connection. There are many error cases that indicate
a logic error in the program -- such as misusing an API or violating a
precondition -- that should be fatal.

Change the exception class for these condition checks to be
std::logical_error, which will then cause fatal exceptions to be thrown.