SMS: make storage errors fatal
Failing a write to disk and a socket getting disconnected are classic
examples of errors that can only be detected at runtime. SMS properly
uses std::runtime_error to indicate these issues, but this fails to
distinguish between errors that should be fatal (writing to disk) and
those that are transient (disconnected socket).

As DataSource drives most activity in SMS in response to new data
arriving, it catches runtime errors, resets the connection, and tries to
reconnect. This works well when the source of the error is a transient,
but we would prefer to exit SMS when we are no longer able to storage
data reliably.

While the proper solution would be to change the exception hierarchy to
distinguish these error cases, we will convert write errors to use
std::logic_error for now. This will cause them to become fatal to SMS,
and is not too far of a stretch, as the code is written such that they
should never occur absent logic bugs in SMS or outside interference in
the operating environment.