SMS: improve marker handling
Live clients need to know if we're currently in a scan, and if so, what
the current scan index is. Adding this functionality properly requires
that we maintain some internal state about the scan index and comments
to avoid a race with the control system configuring its next annotation.

Similarly, live clients will need to know to stop any outstanding scan
when we start a new run.