Parser: refactor POSIX requirements out of the base parser
Using POSIX in the basic parser makes it more difficult to use the class
under Windows, or under other scenarios where we may not be getting the
stream from a file descriptor. There are really two parts to the
original Parser design -- get the data into a buffer, and parse the
buffer. We split the two activities by exposing the information needed
to fill the Parser's buffer to child classes, and giving the child class
a way to initiate processing.

We then create a new POSIXParser class that provides the ability to read
from a file descriptor into the Parser's internal buffer. This satisfies
the needs of the existing clients, and gives Windows users a path
forward.