SMS: periodically snapshot our state
In preparation for replaying the stream from the last run transition,
or up to the current time, we need to periodically take a snapshot of
the experimental state.

Whenever the current pulse time exceeds the index period, we will create
a new snapshot, and push it to the front of our state index. This keeps
the order of timestamps decending, which will make it easier to find the
first snapshot before the requested time when we replay.

We also take the opportunity to index the snapshot we already take when
creating a new file for the container, as this only costs us a small
amount of memory to store the entry -- the state is already in the data
file.