SMS: allow live clients to request replay of stream data
Now that we can recreate the experiment state at periodic snapshots, we
can support replay back to the last run transition, and at points in
between.

When looking for (recent) historical data, we iterate the state index
looking for the first entry that is older than our requested timestamp;
that becomes the starting point for replay. We then inform the caller of
each segment of data that needs to be streamed out, up to the currently
active file.