SMS: exit cleanly if we don't have a log appender
If we didn't have a log appender configured -- ie, no config file -- we
would crash while trying to clean up the ConsoleAppender we created.
Much better to just cleanly exit and be done with it.

We don't need to destroy the ConsoleAppender object we create, as it
appears that log4cxx will handle that for us after calling
removeAppender() on the root logger. Trying to delete the object
ourselves results in a double free error from glibc.