SMS: split StorageManager initialization
We want to handle part of the storage manager initialization later in
the startup sequence; if we create the initial container too early, the
Geometry and PixelMap classes will not have been instantiated yet, so
those packets will not be in the initial file.

This also moves the background thread creation to the late init --
paving the way for proper ownership once we become a daemon.