SMS: don't leak file descriptors
We were leaking file descriptors from the state files we created. This
was a fairly slow leak -- we only made a state file on live client
connection -- but now that we make periodic snapshots, we leak enough to
cause other problems pretty quickly.

EPICS uses select(), and limits the maximum fd value to 1023. Once we
have leaked enough file descriptors that new ones are above that value,
we stop being able to send data to STS until we restart the daemon.

The fix is simple; release our reference on the state file once we have
filled it with data. As the file knows its own name, it can be reopened
as needed by clients.