SMS: add TDC offset to packet timestamp for choppers
To ensure each chopper TDC gets a reasonably unique packet timestamp, we
add the time-of-flight information to the pulse id. This allows multiple
TDC events for a chopper per pulse, and avoids downstream clients
loosing the events.

This transformation could be handled individually by the clients, but
will be handled here to ensure consistency.