doc: match BeamMonitor packet semantics to deployed practice
The TOF values in the BeamMonitor packet were specified such that bit 31
set indicated the falling edge. This is opposite to the way the raw
event field IDs are allocated and contrary to use in production at
HYSPEC. As no one is currently inspecting that field, update the
documentation to match our deployed implementation.