build: allow disabling the build of components
It is useful to allow compilation of a subset of the ADARA code when
working in a limited environment -- for example, building STS and DASMON
require packages to be present that usually are not on the SMS machines
servicing the beam lines.

Now that we use autoconf to probe for more of the required packages, we
can automatically disable the build of components when their
dependencies are not met. This will improve as we have autoconf verify
more of the dependencies.