SMS: allocate unique device ids for the output stream
Now that the groundwork is in place, flesh out the implementaion of the
remapping engine. We assign a unique number to each device for the
output stream, always incrementing. Just in case we add/remove 4 billion
devices in a single invocation of SMS, we verify that the selected id is
not currently in use.

Whenever we drop a source, we will need to remove its mapped device
ids from our table.