SMS: groundwork for device remapping
We need to be able to remap device ids to handle conflicting values when
running with multiple sources. Move the copy of the packet to be stored
up the call chain so that we can access the proper methods in their
classes to handle the remapping.

We add an identity remapping function to preserve current behaviour, and
an allocation function that does nothing to ensure uniqueness.