ADARAPackets: allow mutating the device ID for variables
We need to remap the device Id for each variable source, and this is
best handled through the provided classes.

This method is a pretty ugly choice, as we have to cast away const'ness
to make the modification and this is not reflected in the API. The
remapDevice() method should only be called on objects that we know we
own the memory for.